+++ BlackDesertOnline_Codes_Discord/config.json
@@ -1,4 +1,4 @@
 ﻿{
   "token": "MTEwODc1NDEyMTIyMzMxOTY0Mw.GRbM7c.U_FQLE1uP2A7BL0ER4fy0T6JIr_RlLOpl2-2hU",
-  "prefix":  "*"
+  "prefix": "*"
 }
